+++ crates/duckdb/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Configure s3 credential chain ([#621](https://github.com/stac-utils/stac-rs/pull/621))
 - Read hive partitioned datasets, `Config` structure ([#624](https://github.com/stac-utils/stac-rs/pull/624))
+- `Client.search_to_arrow_table` ([#634](https://github.com/stac-utils/stac-rs/pull/634))
 
 ## [0.1.1] - 2025-01-31